letsrobot/requirements.txt
PySerial
zerorpc
